\name{recommenderlabrats-package}
\alias{recommenderlabrats-package}
\alias{recommenderlabrats}
\docType{package}
\title{
This package implements some practical recommendation engines
}
\description{
Currently implemented -
Low Rank Matrix Factorization
}
\details{
\tabular{ll}{
Package: \tab recommenderlabrats\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-10\cr
License: \tab GPL (>= 2)\cr
}
Follow the vignette of recommenderlab to see how to test.
A walkthrough is also available on sanealytics.com
}
\author{
Saurabh Bhatnagar

Maintainer: Saurabh Bhatnagar <Saurabh.Bhatnagar@gmail.com>
}
\references{
Simon Funk
Koren et all
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
